backend/requirements.txt
django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
